[fix]api-10 commonとの連携ビルドエラー
パッケージの切り方を変更
