--- cucp_api/src/main/java/biz/uoray/api/CucpApiApplication.java
@@ -1,14 +1,0 @@
-package biz.uoray.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-public class CucpApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CucpApiApplication.class, args);
-	}
-
-}
