--- cucp_api/src/main/java/biz/uoray/api/common/entity/Car.java
@@ -1,77 +1,0 @@
-package biz.uoray.api.common.entity;
-
-import biz.uoray.common.entity.AbstractEntity;
-import lombok.Data;
-
-import javax.persistence.*;
-import java.util.Date;
-import java.util.List;
-
-@Entity
-@Data
-@Table(name = "cars")
-public class Car extends AbstractEntity {
-
-    /**
-     * デフォルトコンストラクタ
-     */
-    public Car() {
-    }
-
-    /**
-     * 新規作成・更新用コンストラクタ
-     *
-     * @param name
-     */
-    public Car(String code, String name) {
-        this.code = code;
-        this.name = name;
-    }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name = "code")
-    private String code;
-
-    @Column(name = "name")
-    private String name;
-
-    @Temporal(TemporalType.DATE)
-    @Column(name = "created_at")
-    private Date createdAt;
-
-    @Temporal(TemporalType.DATE)
-    @Column(name = "updated_at")
-    private Date updatedAt;
-
-    @Temporal(TemporalType.DATE)
-    @Column(name = "deleted_at")
-    private Date deletedAt;
-
-//    @OneToMany(mappedBy = "car", fetch = FetchType.LAZY)
-//    private List<CarDetail> carDetailList;
-
-    // Getter
-    @Override
-    public Date getCreatedAt() {
-        return this.createdAt;
-    }
-
-    @Override
-    public Date getUpdatedAt() {
-        return this.updatedAt;
-    }
-
-    // Setter
-    @Override
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    @Override
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
