--- cucp_api/src/main/java/biz/uoray/api/common/entity/Store.java
@@ -1,75 +1,0 @@
-//package biz.uoray.api.common.entity;
-//
-//import biz.uoray.common.entity.AbstractEntity;
-//import biz.uoray.common.entity.CarDetail;
-//import lombok.Data;
-//
-//import javax.persistence.*;
-//import java.util.Date;
-//import java.util.List;
-//
-//@Entity
-//@Data
-//@Table(name = "stores")
-//public class Store extends AbstractEntity {
-//
-//    /**
-//     * デフォルトコンストラクタ
-//     */
-//    public Store() {
-//    }
-//
-//    /**
-//     * 新規登録、更新用のコンストラクタ
-//     *
-//     * @param name
-//     */
-//    public Store(String name) {
-//        this.name = name;
-//    }
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Integer id;
-//
-//    @Column(name = "name")
-//    private String name;
-//
-//    @Temporal(TemporalType.DATE)
-//    @Column(name = "created_at")
-//    private Date createdAt;
-//
-//    @Temporal(TemporalType.DATE)
-//    @Column(name = "updated_at")
-//    private Date updatedAt;
-//
-//    @Temporal(TemporalType.DATE)
-//    @Column(name = "deleted_at")
-//    private Date deletedAt;
-//
-//    @OneToMany(mappedBy = "store", fetch = FetchType.LAZY)
-//    private List<CarDetail> carDetailList;
-//
-//    // Getter
-//    @Override
-//    public Date getCreatedAt() {
-//        return this.createdAt;
-//    }
-//
-//    @Override
-//    public Date getUpdatedAt() {
-//        return this.updatedAt;
-//    }
-//
-//    // Setter
-//    @Override
-//    public void setCreatedAt(Date createdAt) {
-//        this.createdAt = createdAt;
-//    }
-//
-//    @Override
-//    public void setUpdatedAt(Date updatedAt) {
-//        this.updatedAt = updatedAt;
-//    }
-//
-//}
