--- cucp_api/src/main/java/biz/uoray/api/common/repository/CarDetailRepository.java
@@ -1,9 +1,0 @@
-//package biz.uoray.api.common.repository;
-//
-//import biz.uoray.common.entity.CarDetail;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface CarDetailRepository extends JpaRepository<CarDetail, Integer> {
-//}
