--- cucp_api/src/main/java/biz/uoray/api/common/repository/CarRepository.java
@@ -1,9 +1,0 @@
-package biz.uoray.api.common.repository;
-
-import biz.uoray.api.common.entity.Car;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CarRepository extends JpaRepository<Car, Integer> {
-}
