--- cucp_api/src/main/java/biz/uoray/api/common/repository/PriceRepository.java
@@ -1,9 +1,0 @@
-//package biz.uoray.api.common.repository;
-//
-//import biz.uoray.common.entity.Price;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface PriceRepository extends JpaRepository<Price, Integer> {
-//}
