--- cucp_api/src/main/java/biz/uoray/api/common/repository/StoreRepository.java
@@ -1,17 +1,0 @@
-//package biz.uoray.api.common.repository;
-//
-//import biz.uoray.common.entity.Store;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface StoreRepository extends JpaRepository<Store, Integer> {
-//
-//    /**
-//     * storeテーブルを販売店名でselect
-//     *
-//     * @param name
-//     * @return
-//     */
-//    public Store findByName(String name);
-//}
