--- cucp_api/src/main/java/biz/uoray/api/common/repository/custom/GraphRepositoryCustom.java
@@ -1,25 +1,0 @@
-//package biz.uoray.api.common.repository.custom;
-//
-//import biz.uoray.common.entity.Car;
-//import biz.uoray.common.entity.CarDetail;
-//
-//import java.util.List;
-//
-//public interface GraphRepositoryCustom {
-//
-//
-//    /**
-//     * carsの一覧を取得
-//     *
-//     * @return
-//     */
-//    public List<Car> findAll();
-//
-//    /**
-//     * car_detailsの一覧を取得
-//     *
-//     * @param carId
-//     */
-//    public List<CarDetail> findByCarId(int carId);
-//
-//}
