--- cucp_api/src/main/java/biz/uoray/api/common/validation/PositiveDouble.java
@@ -1,37 +1,0 @@
-package biz.uoray.api.common.validation;
-
-import biz.uoray.api.common.validation.validator.PositiveDoubleValidator;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import javax.validation.ReportAsSingleViolation;
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-/**
- * 0.0以上の小数であるかどうか
- */
-@Documented
-@Constraint(validatedBy = {PositiveDoubleValidator.class})
-@Target({METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER})
-@Retention(RUNTIME)
-@ReportAsSingleViolation
-public @interface PositiveDouble {
-
-    String message() default "正の小数を入力してください";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-
-    @Target({METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER})
-    @Retention(RUNTIME)
-    @Documented
-    @interface List {
-        PositiveDouble[] value();
-    }
-}
