--- cucp_api/src/main/java/biz/uoray/api/dto/GraphDto.java
@@ -1,28 +1,0 @@
-//package biz.uoray.api.dto;
-//
-//import java.util.List;
-//import java.util.Map;
-//
-//public class GraphDto {
-//
-//    private List<String> dateLabelList;
-//
-//    Map<String, String[]> priceLabelMap;
-//
-//    public List<String> getDateLabelList() {
-//        return dateLabelList;
-//    }
-//
-//    public void setDateLabelList(List<String> dateLabelList) {
-//        this.dateLabelList = dateLabelList;
-//    }
-//
-//    public Map<String, String[]> getPriceLabelMap() {
-//        return priceLabelMap;
-//    }
-//
-//    public void setPriceLabelMap(Map<String, String[]> priceLabelMap) {
-//        this.priceLabelMap = priceLabelMap;
-//    }
-//
-//}
