+++ cucp_api/src/main/java/biz/uoray/cucp/CucpApiApplication.java
@@ -1,0 +1,13 @@
+package biz.uoray.cucp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CucpApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CucpApiApplication.class, args);
+    }
+
+}
