+++ cucp_api/src/main/java/biz/uoray/cucp/config/WebConfig.java
@@ -1,0 +1,21 @@
+package biz.uoray.cucp.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer{
+
+	@Override
+	public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+		PageableHandlerMethodArgumentResolver resolver = new PageableHandlerMethodArgumentResolver();
+		resolver.setOneIndexedParameters(true);
+		resolver.setMaxPageSize(5);
+		argumentResolvers.add(resolver);
+	}
+
+}
