--- cucp_api/src/test/java/biz/uoray/api/ApiApplicationTests.java
@@ -1,13 +1,0 @@
-package biz.uoray.api;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
