+++ cucp_api/src/test/java/biz/uoray/cucp/ApiApplicationTests.java
@@ -1,0 +1,13 @@
+package biz.uoray.cucp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
